% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm_delay_easy.R
\name{ccm_delay_easy}
\alias{ccm_delay_easy}
\title{ccm_delay_easy}
\usage{
ccm_delay_easy(df, best_E, lib_segments, pred_segments = lib_segments,
  tp = -3:3, ...)
}
\arguments{
\item{df}{A data.frame where each column is a time series. Note that additional columns such as time and replicates should not be included in this data.frame.}

\item{best_E}{The embedding dimensions to use for time delay embedding.}

\item{lib_segments}{Segments which are used for predictions.}

\item{pred_segments}{Segments which are used for predictions.}

\item{tp}{Range of prediction lag should be tested. Default is -10 to 10.}

\item{...}{Additional arguments to be passed to ccm.}
}
\value{
A data.frame with forecast statistics for each prediction lag tested.
}
\description{
Performs convergent cross mapping using simplex projection across all pairwise comparisons and a range of prediction lag.
}
\keyword{ccm}
\keyword{rEDM,}
