% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm_delay_plot.R
\name{ccm_delay_plot}
\alias{ccm_delay_plot}
\title{ccm_delay_plot}
\usage{
ccm_delay_plot(summaryout)
}
\arguments{
\item{summaryout}{output of a call "ccm_summary".}

\item{...}{Additional arguments to be passed to ggplot2.}
}
\value{
A ggplot2 object.
}
\description{
Plot prediction skills as a function of prediction lag (tp).
}
\keyword{ggplot2}
\keyword{rEDM,}
