% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm_easy.R
\name{ccm_easy}
\alias{ccm_easy}
\title{ccm_easy}
\usage{
ccm_easy(df, best_E, lib_segments, pred_segments = lib_segments,
  vars_matrix_ord = NULL, ...)
}
\arguments{
\item{df}{A data.frame where each column is a time series. Note that additional columns such as time and replicates should not be included in this data.frame.}

\item{best_E}{The embedding dimensions to use for time delay embedding.}

\item{lib_segments}{Segments which are used for predictions.}

\item{pred_segments}{Segments which are used for predictions}

\item{vars_matrix_ord}{An optional pre-made data.frame of combinations of variables to test. Must include columns 'lib_column', 'target_column', and 'best_E'.}

\item{...}{Additional arguments to be passed to ccm}
}
\value{
A data.frame with forecast statistics
}
\description{
Performs convergent cross mapping using simplex projection across all pairwise comparisons.
}
\keyword{ccm}
\keyword{rEDM,}
