% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm_easy_plot.R
\name{ccm_easy_plot}
\alias{ccm_easy_plot}
\title{ccm_easy_plot}
\usage{
ccm_easy_plot(summaryout)
}
\arguments{
\item{summaryout}{output of a call "ccm_easy".}

\item{...}{Additional arguments to be passed to ggplot2.}
}
\value{
A ggplot2 object
}
\description{
Plots convergent cross mapping results.
}
\keyword{ggplot2}
\keyword{rEDM,}
