% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_E.R
\name{find_max_E}
\alias{find_max_E}
\title{find_max_E}
\usage{
find_max_E(simplex_out, df)
}
\arguments{
\item{simplex_out}{output of a call from "simplex"}

\item{df}{a data frame that includes columns "E" (embedding dimension) and "rho" (prediction skill)}
}
\value{
Value of E corresponding to the best predictive skill.
}
\description{
Finds the embedding dimension corresponding to the best predictive skill for a simplex call.
}
\keyword{dimension}
\keyword{embedding}
\keyword{rEDM,}
