% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_theta.R
\name{find_max_theta}
\alias{find_max_theta}
\title{find_max_theta}
\usage{
find_max_theta(smap_out, buffer = 0.99, predtype = "rho")
}
\arguments{
\item{smap_out}{output of a call from "smap"}

\item{buffer}{A number from 0 to 1, indicating the allowed buffer distance between the "best" E, and the E that is used. If mulple E values fall within the buffer, then the smallest E is used. Defaults to 0.99, i.e. any value within 99\% of the "best" E can be chosen.}

\item{predtype}{A character indicating which metric to be used for determining predictive skill. Defaults to "rho". Can be "rho", "mae", or "rmse".}
}
\value{
Value of E corresponding to the best predictive skill.
}
\description{
Finds the embedding dimension corresponding to the best predictive skill for a simplex call.
}
\keyword{dimension}
\keyword{embedding}
\keyword{rEDM,}
