% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tp.R
\name{find_tp}
\alias{find_tp}
\title{find_tp}
\usage{
find_tp(sgn, cutoff = 0.05)
}
\arguments{
\item{sgn}{output of a call from "ccm_significance".}

\item{cutoff}{cuttoff for p-value. Default is 0.05.}
}
\value{
Data frame with the tps corresponding to the best predictive skill.
}
\description{
Finds the prediction lag tp corresponding to a convergent cross mapping significance test.
}
\keyword{lag}
\keyword{prediction}
\keyword{rEDM,}
