% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smap_coef.R
\name{get_smap_coef}
\alias{get_smap_coef}
\title{get_smap_coef}
\usage{
get_smap_coef(df, lib_segments, sigout, best_E, best_theta, ...)
}
\arguments{
\item{df}{A data.frame where each column is a time series. Note that additional columns such as time and replicates should not be included in this data.frame.}

\item{lib_segments}{Segments which are used for predictions.}

\item{sigout}{results of ccm significance test, from significance_test() or ccm_summary() functions}

\item{best_E}{vector of best E values from find_max_E()}

\item{best_theta}{vector of best theta values from find_max_theta()}

\item{...}{Additional arguments to be passed to block_lnlp - e.g. if different pred segments are desired}
}
\value{
List including the direction of tets, list of s-mapping results, and list of blocks of data corresponding to the tests.
}
\description{
Calculate s-mapping coefficients based on results from ccm-test
}
\keyword{rEDM,}
\keyword{s-mapping}
