% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{inormalize_data}
\alias{inormalize_data}
\title{inormalize_data}
\usage{
inormalize_data(normout, x = NA)
}
\arguments{
\item{normout}{Output from normalize_data function.}

\item{x}{An optional vector of values to be untranformed based on the mean and standard deviation stored in normalize_data. Defaults to NA, in which case the values in normout are used.}
}
\value{
Untransformed values
}
\description{
Inverse transformation of normalize_data function.
}
\keyword{normalization}
