% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(x, truezero = FALSE)
}
\arguments{
\item{x}{a numeric vector to be normalized}

\item{truezero}{A logical telling whether the data includes true zeros - if TRUE then standardizes based only on the standard deviaiton of x, else if FALSE standardizes x based on its mean and standard deviation. Defaults to FALSE.}
}
\value{
A list, including the standardized values of x (xnorm), and the mean (mux) and the standard deviation (sdx) used for the transformation.
}
\description{
Normalizes data, either for data with or without true zeros.
}
\keyword{normalization}
