% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simplex.R
\name{plot_simplex}
\alias{plot_simplex}
\title{plot_simplex}
\usage{
plot_simplex(simplex_out, predtype = "rho", pname = NA)
}
\arguments{
\item{simplex_out}{Output from simplex function (or a list of outputs)}

\item{predtype}{Type of prediction metric to use. Can be "rho", "rmse", or "mae".}

\item{pname}{Optional name for the plot}
}
\value{
A ggplot object
}
\description{
Plot prediction skill vs. theta for the results of the simplex function
}
\keyword{rEDM,}
\keyword{s-map}
