% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smap.R
\name{plot_smap}
\alias{plot_smap}
\title{plot_smap}
\usage{
plot_smap(smap_out, predtype = "rho", pname = NA)
}
\arguments{
\item{smap_out}{Output from s-map function (or a list of outputs).}

\item{predtype}{Type of prediction metric to use. Can be "rho", "rmse", or "mae".}

\item{pname}{Optional name for the plot.}
}
\value{
A ggplot object.
}
\description{
Plot prediction skill vs. theta for the results of an s-map.
}
\keyword{rEDM,}
\keyword{s-map}
