% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_data.R
\name{segment_data}
\alias{segment_data}
\title{segment_data}
\usage{
segment_data(replicates)
}
\arguments{
\item{replicates}{A data column contains the replicates to be splitted}
}
\value{
A matrix including the beginning and the end of the segments
}
\description{
Splits data by replicates, create segments for predictions
}
\examples{
fakedata = cbind(rep(1:3, each=10), rnorm(30))
segment_data(fakedata[,1])
}
\keyword{replicates}
\keyword{split,}
