% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm_significance.R
\name{significance_test}
\alias{significance_test}
\title{significance_test}
\usage{
significance_test(ccm_output, predtype = "rho")
}
\arguments{
\item{ccm_output}{output of a call from "ccm_easy"}

\item{predtype}{A character indicating which metric to be used for determining predictive skill. Defaults to "rho". Can be "rho", "mae", or "rmse".}
}
\value{
Results from significance test.
}
\description{
Significance test for convergent cross mapping
}
\keyword{rEDM,}
\keyword{significance}
