% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_coef_plot.R
\name{smap_coef_plot}
\alias{smap_coef_plot}
\title{smap_coef_plot}
\usage{
smap_coef_plot(index_sp, smap_coef_out, covar = "time", ycol = "all",
  xlab = "")
}
\arguments{
\item{index_sp}{Index number for the desired plot - Calls the corresponding analysis number from smap_coef_out.}

\item{smap_coef_out}{output of a call "get_smap_coef".}

\item{covar}{Covariate to plot on x-axis. Defaults to "time". Can be either a column name from the block, or a vector.}

\item{ycol}{S-map coefficient columns to plot. Defaults to 'all' - i.e. all variables.}

\item{xlab}{Name to plot on x-axis. Defaults to "".}

\item{...}{Additional arguments to be passed to ggplot2.}
}
\value{
A ggplot2 object
}
\description{
Plot s-map coefficients.
}
\keyword{ggplot2}
\keyword{rEDM,}
