% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.R
\name{stability_analysis}
\alias{stability_analysis}
\title{stability_analysis}
\usage{
stability_analysis(smap_coef_out, doplot = TRUE)
}
\arguments{
\item{smap_coef_out}{Output of a call "get_smap_coef".}

\item{doplot}{Logical argument indicating if indices should be plotted or not. Default is TRUE.}
}
\value{
List including indices, and plots if requested.
}
\description{
Calculate local stability and interaction indices.
}
\keyword{rEDM,}
\keyword{s-mapping}
